Tehtävät 4.18. - 4.22 + tehtävät 4.6. - 4.7.

+++ Osa4/blogilista_backend/controllers/blogs.js
@@ -1,12 +1,21 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+const helper = require('../utils/list_helper')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog.find({}).populate('user', { userName: 1, name: 1 })
+  // Extra: Calculate statistics using helper functions
+  const mostBlogs = helper.mostBlogs(blogs)
+  const favoriteBlog = helper.favoriteBlog(blogs)
+  const mostLikes = helper.mostLikes(blogs)
+  response.json({ blogs, mostBlogs, favoriteBlog, mostLikes })
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const user = request.user
   const body = request.body
 
   const blog = new Blog({
@@ -14,9 +23,13 @@ blogsRouter.post('/', async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
@@ -29,9 +42,16 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'User not authorized to delete this blog' })
+  } else {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -41,7 +61,7 @@ blogsRouter.put('/:id', async (request, response) => {
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
   }
 
   const blogtoUpdate = await Blog.findById(request.params.id)
